.github/scripts/requirements.txt
# Python dependencies for governance scripts
pyyaml>=6.0
